# services/auction-dutch/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation",
        "title": "Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://aban.io/support",
            "email": "info@aban.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/apis/auctions/{id}": {
            "get": {
                "description": "Public read: the Dutch auction's state plus the SERVER-authoritative current price at now and the next drop instant. Clients render from these; the buy decision is always re-validated server-side.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction-dutch"
                ],
                "summary": "Auction state + live price",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auction UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AuctionResp"
                        }
                    },
                    "400": {
                        "description": "Bad UUID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AuctionResp": {
            "type": "object",
            "properties": {
                "ceilingCents": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "currentPriceCents": {
                    "description": "server-authoritative price(now)",
                    "type": "integer"
                },
                "dropIntervalSeconds": {
                    "type": "integer"
                },
                "dropStepCents": {
                    "type": "integer"
                },
                "floorCents": {
                    "type": "integer"
                },
                "hammerAt": {
                    "type": "string"
                },
                "hammerPriceCents": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "lotId": {
                    "type": "string"
                },
                "nextDropAt": {
                    "type": "string"
                },
                "openAt": {
                    "type": "string"
                },
                "state": {
                    "description": "DRAFT | ... | OPEN | HAMMER | ...",
                    "type": "string"
                },
                "winnerAccountId": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}